storage_node/requirements.txt
flask==2.2.3
requests==2.28.2
gunicorn==20.1.0
python-dotenv==1.0.0
flask-cors==3.0.10
pycryptodome==3.17.0 